Re-factor BlueZ manager classes to reduce code duplication
