+++ blueman/bluez/Makefile.am
@@ -11,6 +11,7 @@ blueman_PYTHON =				\
 	Device.py					\
 	errors.py					\
 	Manager.py					\
+	ManagerBase.py					\
 	PropertiesBase.py				\
 	Network.py					\
 	NetworkServer.py
